feature/product-resources

+++ src/product/product.service.ts
@@ -1,3 +1,4 @@
+import { ResponseError } from '../error/response-error';
 import {
     GetProductResponse,
     PostProductBody,
@@ -25,11 +26,11 @@ export class ProductService {
         // Validate request with zod
         const requestValidated = PostProductVal.parse(body);
         // Split request into product and productItem
-        const { productItem, ...product } = requestValidated;
+        const { productItems, ...products } = requestValidated;
         // Insert product
-        const response = await ProductRepository.postProduct(product);
+        const response = await ProductRepository.postProduct(products);
         // Map productId to productItem
-        const productItemWithProductId = productItem.map((item) => ({
+        const productItemWithProductId = productItems.map((item) => ({
             productId: response.id,
             ...item,
         }));
@@ -39,7 +40,33 @@ export class ProductService {
         return { id: response.id };
     }
 
+    static async postProductImage(file: Express.Multer.File, param: ProductIdParam): Promise<PostProductResponse> {
+        // Check productId
+        const check = await ProductRepository.getProductById(param);
+        if (!check) {
+            throw new ResponseError(404, 'Product not found', 'Produk tidak ditemukan');
+        }
+
+        // Validate file
+        if (!file) {
+            throw new ResponseError(400, 'File is required', 'File tidak boleh kosong');
+        }
+        const input = {
+            image: file.filename,
+        };
+
+        // Update table product
+        const result = await ProductRepository.postProductImage(input, param);
+
+        return result;
+    }
+
     static async putProduct(body: PutProductBody, param: ProductIdParam): Promise<PutProductResponse> {
+        // Check productId
+        const check = await ProductRepository.getProductById(param);
+        if (!check) {
+            throw new ResponseError(404, 'Product not found', 'Produk tidak ditemukan');
+        }
         // Validate request with zod
         const requestValidated = PutProductVal.parse(body);
         // Update table product
@@ -49,6 +76,11 @@ export class ProductService {
     }
 
     static async deleteProduct(param: ProductIdParam) {
+        // Check productId
+        const check = await ProductRepository.getProductById(param);
+        if (!check) {
+            throw new ResponseError(404, 'Product not found', 'Produk tidak ditemukan');
+        }
         // Update table product
         await ProductRepository.deleteProduct(param);
     }
@@ -57,6 +89,11 @@ export class ProductService {
         body: PutProductItemBody,
         param: ProductIdParam & ProductItemIdParam,
     ): Promise<PutProductItemResponse> {
+        // Check productId
+        const check = await ProductRepository.getProductById(param);
+        if (!check) {
+            throw new ResponseError(404, 'Product not found', 'Produk tidak ditemukan');
+        }
         // Validate request with zod
         const requestValidated = PutProductItemVal.parse(body);
         // Update table productItem
@@ -66,6 +103,11 @@ export class ProductService {
     }
 
     static async deleteProductitem(param: ProductIdParam & ProductItemIdParam) {
+        // Check productId
+        const check = await ProductRepository.getProductById(param);
+        if (!check) {
+            throw new ResponseError(404, 'Product not found', 'Produk tidak ditemukan');
+        }
         // Update table productItem
         await ProductRepository.deleteProductItem(param);
     }
